render: Calculate correct bounds for shapes containing curves
